✨ Split mocks/handlers.ts (2474 → 2560 lines, better organized): fixtures (400L) + endpoints (2140L) + re-export barrel (22L) — #10439
